backend migrated: SQLite database into PostgreSQL database and generalized ORM

+++ tk_water_backend/WebApplication1/WebApplication1/Controllers/backendBodys/receiveAllUnitsData/ReceiveAllUnitsDataRequestBody.cs
@@ -2,6 +2,6 @@
 {
     public class ReceiveAllUnitsDataRequestBody
     {
-        public long UserID { get; set; }
+        public int UserID { get; set; }
     }
 }
