+++ tk_water_backend/WebApplication1/WebApplication1/Controllers/controlCentrumBodys/postUnitMeasurement/PostUnitMeasurementRequestBody.cs
@@ -3,6 +3,6 @@
     public class PostUnitMeasurementRequestBody
     {
         public string UnitID { get; set; } = string.Empty;
-        public long MoistureLevel { get; set; }
+        public short MoistureLevel { get; set; }
     }
 }
