+++ tk_water_backend/WebApplication1/WebApplication1/Controllers/controlCentrumBodys/registerUnit/RegisterUnitRequestBody.cs
@@ -5,8 +5,8 @@ namespace WebApplication1.Controllers.controlCentrumBodys.registerUnit
     public class RegisterUnitRequestBody
     {
         public string UnitID { get; set; } = string.Empty;
-        public long MoistureLevel { get; set; }
-        public long MoistureThreshold { get; set; }
+        public short MoistureLevel { get; set; }
+        public short MoistureThreshold { get; set; }
 
     }
 }
